Revert "Added contacts and home page"
