--- client/src/App.css
@@ -1,20 +1,0 @@
-/* Basic styling for the app */
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    background-color: #fff;
-    color: #333;
-  }
-  
-  .container {
-    width: 100%;
-    padding: 20px;
-  }
-  
-  h1 {
-    font-size: 2rem;
-    color: #333;
-  }
-
