+++ client/src/App.tsx
@@ -1,18 +1,17 @@
-import './App.css'; // Make sure this line is included at the top
-import Navbar from './components/Navbar';
-import SideNav from './components/SideNav';
 import { Outlet } from 'react-router-dom';
 
+import Navbar from './components/Navbar';
+
 function App() {
+
   return (
     <div>
       <Navbar />
-      <SideNav />
-      <main className="container pt-5">
+      <main className='container pt-5'>
         <Outlet />
       </main>
     </div>
-  );
+  )
 }
 
-export default App;
+export default App
