--- client/src/api/mainNewsAPI.tsx
@@ -1,25 +1,0 @@
-import Auth from '../utils/auth';
-
-const retrieveNews = async () => {
-  try {
-    const response = await fetch('/api/news', {
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${Auth.getToken()}`
-      }
-    });
-    const data = await response.json();
-
-    if(!response.ok) {
-      throw new Error('Invalid user API response, check network tab!');
-    }
-
-    return data;
-
-  } catch (err) { 
-    console.log('Error from data retrieval:', err);
-    return [];
-  }
-}
-
-export { retrieveNews };
