--- client/src/components/NewsComponent.tsx
@@ -1,52 +1,0 @@
-import { useEffect, useState } from "react";
-
-interface NewsSource {
-  id: string | null;
-  name: string;
-}
-
-interface Article {
-  source: NewsSource;
-  author: string | null;
-  title: string;
-  description: string;
-  url: string;
-  urlToImage: string | null;
-  publishedAt: string;
-  content: string | null;
-}
-
-const API_URL = "http://localhost:3001/news";
-
-const NewsComponent = () => {
-  const [articles, setArticles] = useState<Article[]>([]);
-
-  useEffect(() => {
-    const fetchNews = async () => {
-      try {
-        const response = await fetch(API_URL);
-        const data = await response.json();
-        setArticles(data.articles);
-      } catch (error) {
-        console.error("Error fetching news:", error);
-      }
-    };
-
-    fetchNews();
-  }, []);
-
-  return (
-    <div>
-      <h1>Latest News</h1>
-      {articles.map((article, index) => (
-        <div key={index}>
-          <h3>{article.title}</h3>
-          <p>{article.description}</p>
-          <a href={article.url} target="_blank">Read More</a>
-        </div>
-      ))}
-    </div>
-  );
-};
-
-export default NewsComponent;
