--- client/src/components/SideNav.css
@@ -1,78 +1,0 @@
-/* Side navigation bar styles */
-.sidenav {
-    background-color: #fff;
-    position: fixed;
-    top: 30;
-    left: 0;
-    height: 100%;
-    width: 200px;
-    padding: 1rem;
-    box-shadow: 2px 0px 10px rgba(0, 0, 0, 0.1);
-    transition: width 0.3s ease;
-  }
-  
-  .sidenav.closed {
-    width: 60px;
-  }
-  
-  .sidenav-toggle {
-    background-color: #6dc2ff;
-    border: none;
-    padding: 10px;
-    color: white;
-    cursor: pointer;
-    border-radius: 50%;
-    position: absolute;
-    top: 20px;
-    left: -30px;
-  }
-  
-  .sidenav-links {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    margin-top: 50px;
-  }
-  
-  .sidenav-link {
-    display: flex;
-    align-items: center;
-    color: #000;
-    font-size: 1.1rem;
-    text-decoration: none;
-    font-weight: 600;
-    transition: background-color 0.3s ease;
-  }
-  
-  .sidenav-link:hover {
-    background-color: #f0f0f0;
-    padding-left: 10px;
-    border-radius: 5px;
-  }
-  
-  .sidenav-icon {
-    margin-right: 10px;
-    font-size: 1.5rem;
-  }
-  
-  .sidenav.hide {
-    display: none;
-  }
-  
-  .sidenav.open {
-    display: block;
-  }
-  
-  @media screen and (max-width: 768px) {
-    .sidenav {
-      width: 100%;
-      position: absolute;
-      z-index: 1000;
-    }
-  
-    .sidenav-links {
-      flex-direction: row;
-      justify-content: space-around;
-    }
-  }
-
