+++ client/src/interfaces/UserLogin.tsx
@@ -2,5 +2,4 @@
 export interface UserLogin {
   username: string | null;  // Property for storing username, nullable
   password: string | null;  // Property for storing password, nullable
-  email?: string | null;
 }
