--- client/src/pages/ChatGpt.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ChatGpt = () => {
-  return (
-    <div>ChatGpt</div>
-  )
-}
-
-export default ChatGpt
