--- client/src/pages/Entertainment.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Entertainment = () => {
-  return (
-    <div>Entertainment</div>
-  )
-}
-
-export default Entertainment
