--- client/src/pages/Gaming.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Gaming = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Gaming
