--- client/src/pages/Puzzle.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Puzzle = () => {
-  return (
-    <div>Puzzle</div>
-  )
-}
-
-export default Puzzle
