--- client/src/pages/Sports.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Sports = () => {
-  return (
-    <div>Sports</div>
-  )
-}
-
-export default Sports
