--- client/src/pages/Travel.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Travel = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Travel
