--- client/src/pages/Trending.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Trending = () => {
-  return (
-    <div>Trending</div>
-  )
-}
-
-export default Trending
