+++ client/vite.config.ts
@@ -1,9 +1,7 @@
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import path from 'path';
 
+// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
   server: {
     port: 3000,
     open: true,
@@ -16,17 +14,8 @@ export default defineConfig({
       '/auth': {
         target: 'http://localhost:3001',
         changeOrigin: true,
-        secure: false,
+        secure: false
       },
-    },
-    // 👇 This is the real fix
-    fs: {
-      strict: false,
-    },
-  },
-  resolve: {
-    alias: {
-      '@': path.resolve(__dirname, 'src'),
     },
   },
 });
