+++ package.json
@@ -20,22 +20,11 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/react-icons": "^2.2.7",
     "concurrently": "^8.2.2",
     "nodemon": "^3.1.3",
     "wait-on": "^7.2.0"
   },
   "dependencies": {
-    "axios": "^1.8.4",
-    "dotenv": "^16.4.7",
-    "express": "^4.21.2",
-    "framer-motion": "^12.6.2",
-    "jsonwebtoken": "^9.0.2",
-    "main": "file:",
-    "openai": "^4.89.0",
-    "pg": "^8.14.1",
-    "react-icons": "^5.5.0",
-    "react-router-dom": "^7.4.0",
-    "sequelize": "^6.37.6"
+    "bootstrap": "^5.3.3"
   }
 }
