+++ server/db/schema.sql
@@ -1,5 +1,5 @@
 -- DROP DATABASE
-DROP DATABASE IF EXISTS login_db;
+DROP DATABASE IF EXISTS friends_db;
 
 -- CREATE DATABASE
-CREATE DATABASE login_db;
+CREATE DATABASE friends_db;
