+++ server/src/server.ts
@@ -5,17 +5,18 @@ dotenv.config();
 
 import express from 'express';
 import sequelize from './config/connection.js';
-import routes from './routes/index.js'; 
+import routes from './routes/index.js';
 
 const app = express();
 const PORT = process.env.PORT || 3001;
 
+// Serves static files in the entire client's dist folder
 app.use(express.static('../client/dist'));
 
 app.use(express.json());
-app.use(routes); 
+app.use(routes);
 
-sequelize.sync({ force: forceDatabaseRefresh }).then(() => {
+sequelize.sync({force: forceDatabaseRefresh}).then(() => {
   app.listen(PORT, () => {
     console.log(`Server is listening on port ${PORT}`);
   });
